requirements.txt
fastapi
uvicorn
torch
torchvision
Pillow
python-dotenv
google-cloud-aiplatform
vertexai
google-cloud-translate==2.0.1
google-cloud-texttospeech==2.14.1
